+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New function
 * `tg_channel_avg_posts_reach()` - Getting the average coverage of channel publications over time.
 * `tg_channel_err()` - Obtaining an ERR indicator for a channel in dynamics.
+* `tg_set_check_api_quote()` - Enable and disable API limit rate alerts.
 
 ## New options
 * `tg.check_api_quote` - allows you to disable api quota check.